trying to refactory into separate libraries.  its hard.
