+++ src/PsychicStaticFileHandler.h
@@ -1,0 +1,44 @@
+#ifndef PsychicStaticFileHandler_h
+#define PsychicStaticFileHandler_h
+
+#include "PsychicCore.h"
+#include "PsychicHttpServerRequest.h"
+#include "PsychicHttpServerResponse.h"
+#include "PsychicHttpFileResponse.h"
+
+class PsychicStaticFileHandler {
+  using File = fs::File;
+  using FS = fs::FS;
+  private:
+    bool _getFile(PsychicHttpServerRequest *request);
+    bool _fileExists(const String& path);
+    uint8_t _countBits(const uint8_t value) const;
+  protected:
+    FS _fs;
+    File _file;
+    String _filename;
+    String _uri;
+    String _path;
+    String _default_file;
+    String _cache_control;
+    String _last_modified;
+    bool _isDir;
+    bool _gzipFirst;
+    uint8_t _gzipStats;
+  public:
+    PsychicStaticFileHandler(const char* uri, FS& fs, const char* path, const char* cache_control);
+    bool canHandle(PsychicHttpServerRequest *request);
+    esp_err_t handleRequest(PsychicHttpServerRequest *request);
+    PsychicStaticFileHandler& setIsDir(bool isDir);
+    PsychicStaticFileHandler& setDefaultFile(const char* filename);
+    PsychicStaticFileHandler& setCacheControl(const char* cache_control);
+    PsychicStaticFileHandler& setLastModified(const char* last_modified);
+    PsychicStaticFileHandler& setLastModified(struct tm* last_modified);
+  #ifdef ESP8266
+    PsychicStaticFileHandler& setLastModified(time_t last_modified);
+    PsychicStaticFileHandler& setLastModified(); //sets to current time. Make sure sntp is runing and time is updated
+  #endif
+    //PsychicStaticFileHandler& setTemplateProcessor(AwsTemplateProcessor newCallback) {_callback = newCallback; return *this;}
+};
+
+#endif /* PsychicHttp_h */
